Minor: Extract `BatchCoalescer` to its own module
